requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
discord.py==2.2.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mutagen==1.46.0
prettytable==3.7.0
pycparser==2.21
pycryptodomex==3.17
pydub==0.25.1
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.30.0
soupsieve==2.4.1
urllib3==2.0.2
wcwidth==0.2.6
websockets==11.0.2
yarl==1.9.2
youtube-search==2.1.2
yt-dlp==2023.3.4
